refactor: mudança de validacao para a pasta Middleware
