+++ src/server/routes/Route.ts
@@ -1,7 +1,7 @@
 import { Router } from 'express';
-import { BodyValidator } from '../middleware/Validation';
-import { getById } from '../controllers/GetById';
+import { BodyValidator } from '../middleware/LoginValidation';
+import { loginByEmailAndPassword } from '../controllers/Login';
 
 export const router = Router();
 
-router.post('/login', BodyValidator, getById);
+router.post('/login', BodyValidator, loginByEmailAndPassword);
